requirements.txt
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
defer==1.0.6
dirspec==13.10
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7
django-cors-headers==1.1.0
django-toolbelt==0.0.1
djangorestframework==3.1.3
djangorestframework-expiring-authtoken==0.1.1
dnspython==1.11.1
duplicity==0.6.23
gunicorn==19.3.0
html5lib==0.999
httpie==0.9.2
httplib2==0.8
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
lxml==3.3.3
MySQL-python==1.2.3
oauth==1.0.1
oauthlib==0.6.1
oneconf==0.3.7
PAM==0.4.2
pexpect==3.1
Pillow==2.3.0
piston-mini-client==0.7.5
psycopg2==2.4.5
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
Pygments==2.0.2
pygobject==3.12.0
pyOpenSSL==0.13
pyserial==2.6
pysmbc==1.0.14.1
pysqlite==1.0.1
python-apt===0.9.3.5ubuntu1
python-debian===0.1.21-nmu2ubuntu2
pyxdg==0.25
reportlab==3.0
requests==2.7.0
SecretStorage==2.0.0
sessioninstaller==0.0.0
simplejson==3.3.1
six==1.5.2
software-center-aptd-plugins==0.0.0
static3==0.6.1
system-service==0.1.6
Twisted-Core==13.2.0
Twisted-Web==13.2.0
unity-lens-photos==1.0
urlgrabber==3.9.1
urllib3==1.7.1
virtualenv==1.11.4
wadllib==1.3.2
wheel==0.24.0
xdiagnose===3.6.3build2
yum-metadata-parser==1.1.4
zope.interface==4.0.5
